+++ octobot_channels/producer.py
@@ -104,13 +104,13 @@ class Producer:
         """
         pass
 
-    async def wait_for_processing(self) -> None:
+    async def wait_for_processing(self, **kwargs) -> None:
         """
         Should be used only with SupervisedConsumers
         It will wait until all consumers have notified that their consume() method have ended
         :return: None
         """
-        await asyncio.gather(*[consumer.queue.join() for consumer in self.channel.get_consumers()])
+        await asyncio.gather(*[consumer.queue.join() for consumer in self.channel.get_consumers(**kwargs)])
 
     async def stop(self):
         """