+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.1] - 2019-09-07
+### Added
+- Producer supervised consumer wait method 'wait_for_processing'
+
+### Changed
+- Consumer tests
+
 ## [1.3.0] - 2019-09-07
 ### Added
 - Supervised Consumer that notify the consumption end